Remove not needed file permission change
Setting the permission on the ats_2 file to 0777 is bad since we should not write executable files to /data. With this current functionality timekeep requires invasive sepolicy, such as fowner and fsetid.

Changing the permissions on this file is actually not necessary since timekeep is run as root, and the default owner of ats_2 is root with file permissions of 0600. This is actually sufficient for the timekeep functionality, for both the C and Java parts.

With this patch we can set saner sepolicy for timekeep.

Tested on suzu, confirming the functionality of both timekeep (C) and TimeKeep (Java) services.