TimeKeep: Replace `SystemProperties` private API with `sysprop` definition
Using `android.os.SystemProperties` requires access to Android
private platform APIs which is incompatible with specifying an SDK
version.  Opting in to private APIs is largely unnecessary when the
_vendor-specific_ system property can be accessed through a generated
`sysprop` library.

Unfortunately these libraries are not available to C code, only to
C++, so we'll keep using the hardcoded property name in `timekeep.c`.